% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg17.R
\docType{data}
\name{bg17}
\alias{bg17}
\title{This is the 2017 version of the EJSCREEN dataset plus lat lon and countynames, etc., minus some cols and rows}
\format{data.frame}
\description{
Note the 2018 version of EJSCREEN (released late 2018)
    actually uses ACS2016, which is from 2012-2016 (released late 2017).
  Note the 2019 version of EJSCREEN (released late 2019)
    actually uses ACS2017, which is from 2013-2017 (released late 2018).

 This data set is the EJSCREEN 2017 dataset from the ftp site but with
  fields renamed for easier use in the ejscreen package,
  and some columns dropped (svi6-related, and the 2 alternative versions of an EJ Index)
  and some fields added (lat lon for bg centroids, flagged if any of EJ indexes above 80th percentile in US),
  and state name and state abbrev and county name and FIPS for tract, county, state,
  minus a handful of rows (that had NA values in FIPS? These are left in the bg18 data for now.)
}
\details{
Previously the data.frame was bg not bg17,
  and the file was called bg2017_plus_latlon_etc_minus_some_fields_minus_NA_rows.rdata
}
\concept{datasets}
