% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg20.R
\docType{data}
\name{bg20}
\alias{bg20}
\title{The 2020 version of EJSCREEN data (based on ACS 2014-2018) plus lat lon, countynames, etc., minus some nonessential fields}
\format{
data.frame with 220,333 rows (block groups) and 118 columns in bg20
}
\description{
Note the 2020 version of EJSCREEN (released January 2021 rather than late 2020)
    actually uses ACS2018, which is from 2014-2018 (released by Census late 2019).
  Note the 2019 version of EJSCREEN (released late 2019)
    actually uses ACS2017, which is from 2013-2017 (released by Census late 2018).

  This data set is the EJSCREEN dataset table from the csv file on the ftp site but with
  fields renamed for easier use in the ejscreen package,
  and some columns dropped (text fields about percentiles, used as popup text on maps)
  and state name and state abbrev and county name and FIPS for tract, county, state,
  and lat lon of a point representing each block group
  BUT NOT REMOVING a handful of rows removed from the data (that had NA values in FIPS.ST)
  Also, it does not include the lookup tables of percentiles for USA, Regions, States.
  Also, it does not include the lookup tables of percentiles for USA, Regions, States, which are in the gdb.
}
\details{
Definitions of column names here vs in the gdb or csv on the FTP site files
  are in data(ejscreenformulas)
  and also in a file called EJSCREEN_columns_explained.csv
  in the inst folder of the ejscreen package.
  Technical documentation is at \url{https://www.epa.gov/ejscreen}

  The data was created for this package as follows: \cr\cr
    \code{require(ejscreen); require(ACSdownload)} \cr
    \code{require(proxistat) # for adding the lat lon of each block group }\cr
    \code{require(analyze.stuff); require(ejanalysis); require(readr)} \cr
     \cr
  \code{ # DOWNLOADED AND UNZIPPED  }\cr\cr
  \cr 2020 version of EJSCREEN downloaded 2020-01-27 from public FTP site
  \cr as csv format (zipped)
  \cr then unzipped to csv file
  \cr
  \code{#  bg <- ejscreen.download(folder = getwd(), justreadname = 'EJSCREEN_2020_USPR.csv', addflag = TRUE)} \cr\cr
  # Starts by reading in 141 columns from csv\cr
  # and \link{ejscreen.download} renames fields, drops an ID field, \cr
  # and adds fields called FIPS.TRACT, FIPS.COUNTY, FIPS.ST, countyname, and (optionally) flagged. \cr\cr
  \code{bg <- bg[ , !grepl(pattern = 'pctile\\\\.text', x = names(bg))]} \cr
  \code{bg <- bg[ , names(bg) != 'Shape_Length'] } \cr

  # Then added lat, lon fields for block group centroids, via \link[proxistat]{bg.pts} from proxistat package:\cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lat')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lon')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr\cr
  \code{plot(bg$lon[bg$lon < -50], bg$lat[bg$lon < -50], pch = '.')} \cr\cr
  # \code{sum(is.na(bg$FIPS.ST)) # 13 are NA} \cr

  # THEN USE A NAME SPECIFIC TO THE YEAR: \cr
  \code{bg20 <- bg; rm(bg)} \cr
  \code{save(bg20, file = 'bg20.rdata')} \cr \cr
}
\concept{datasets}
