% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg21.R
\docType{data}
\name{bg21}
\alias{bg21}
\title{EJScreen 2.0 dataset (Feb 2022), plus additional variables, minus some nonessential fields}
\format{
data.frame with 220,333 ? rows (block groups) and 118 ? columns in bg21
}
\description{
\preformatted{
 Any mid/late 2022 update of EJScreen will use ACS 2016-2020 released March 2022.
 
 EJScreen 2.0 is the version released by EPA February 18, 2022 
   (but here it is bg21 because it would have been called the late 2021 version but was delayed)
    bg21 could be renamed given the EJScreen 2.0 name.
    This uses ACS2019, which is from 2015-2019 (released by Census Dec 2020).

  The last version before February 18, 2022 was still the 2020 version.
  Note the 2020 version of EJSCREEN (confusingly released mid 2021 rather than late 2020)
    actually used ACS2018, which was from 2014-2018 (released by Census December 2019). See \link{bg20}

  Note the 2019 version of EJSCREEN (released late 2019)
    actually uses ACS2017, which is from 2013-2017 (released by Census late 2018).

  This data set is the EJSCREEN dataset from the ftp site but with
  fields renamed for easier use in the ejscreen package,
  Columns dropped from ftp version, not included here:
    - text fields about percentiles, used as popup text on maps
    - any obsolete columns related to the 2 alternative versions of an EJ Index phased out (except VNI.eo and VSI.eo are kept)
    - any obsolete columns related to using 6 instead of 2 demographic variables (svi6)
  Columns added here, not in ftp version:
    - lat lon for bg centroids
    - state name and state abbrev, county name, and FIPS for tract, county, state
    - flag indicating if any of EJ indexes above 80th percentile in US

  Detailed info on demog subgroups is in \link{bg21DemographicSubgroups2015to2019}
 }
}
\details{
This will give a quick look at some key stats:
   #round(data.frame(cbind(
   #  subgroups = unlist(ustotals(bg21DemographicSubgroups2015to2019)),
   #  maingroups = unlist(ustotals(bg21[ bg21$ST !='PR', ])))
   #  ), 2)

Note that unlike bg21DemographicSubgroups2015to2019,
   bg21 has PR, but will likely add PR to have it in both.

  Lookup tables are not yet included here: lookup tables of percentiles for USA, Regions, States are on EJSCREEN FTP site.
  For that information, see ejscreen::lookupUSA, lookupRegions, lookupStates, and ejscreen::ejscreen.lookuptables

  Definitions of column names here vs in the gdb or csv on the FTP site files
  are in data(ejscreenformulas)
  and also in a file called EJSCREEN_columns_explained.csv
  in the inst folder of the ejscreen package.
  Technical documentation is at \url{https://www.epa.gov/ejscreen}

  The data was created for this package as follows: \cr\cr
    \code{require(ejscreen); require(ACSdownload)} \cr
    \code{require(proxistat) # for adding the lat lon of each block group }\cr
    \code{require(analyze.stuff); require(ejanalysis); require(readr)} \cr
     \cr
     
  \code{ # IF DOWNLOADED AND UNZIPPED  }\cr\cr
  \cr Latest version of EJSCREEN downloaded from public FTP site
  \cr as csv format (zipped)
  \cr then unzipped to csv file
  \cr
  \code{
  #  bg <- ejscreen.download(folder = getwd(), addflag = TRUE,   \cr
   justreadname = 'EJSCREEN_2021_USPR.csv' )} \cr\cr
  \cr  or  \cr\cr
  
  bg = ejscreen.download()
  
  \cr
  #  \link{ejscreen.download} renames fields, drops an ID field, \cr
  # and adds fields called FIPS.TRACT, FIPS.COUNTY, FIPS.ST, countyname, and (optionally) flagged. \cr\cr
  \code{bg <- bg[ , !grepl(pattern = 'pctile\\\\.text', x = names(bg))]} \cr
  \code{bg <- bg[ , names(bg) != 'Shape_Length'] } \cr

  # Then added lat, lon fields for block group centroids, via \link[proxistat]{bg.pts} from proxistat package:\cr
  # for 2010 boundaries that was like this:\cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lat')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lon')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr\cr
  # for 2020 boundaries it is like this: \cr
  
  
  \code{plot(bg$lon[bg$lon < -50], bg$lat[bg$lon < -50], pch = '.')} \cr\cr
  # \code{sum(is.na(bg$FIPS.ST)) # check if any are NA} \cr

  # THEN USE A NAME SPECIFIC TO THE YEAR: \cr
  \code{bg21 <- bg; rm(bg)} \cr
  \code{usethis::use_data(bg21) #  save(bg21, file = 'bg21.rdata')} \cr \cr
}
\concept{datasets}
