% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg22_ACS2016_2020.R
\docType{data}
\name{bg22_ACS2016_2020}
\alias{bg22_ACS2016_2020}
\title{ACS blockgroup data for EJScreen dataset, from ACS 2016-2020 summary file.}
\format{
data.frame with 220,333 ? rows (block groups) and 118 ? columns
}
\description{
\preformatted{
 Any mid/late 2022 update of EJScreen will use this demographic data,
   from the 5-year summary file American Community Survey ACS 2016-2020
   released March 17 2022.

 EJScreen 2.0 is the version released by EPA February 18, 2022
   (but in this package it is called bg21 because
    it was meant to be the late 2021 version but was delayed)
    bg21 could be renamed given the EJScreen 2.0 name.
    It uses ACS2019, which is from 2015-2019 (released by Census Dec 2020).

  The last version before February 18, 2022 was still the 2020 version.
  Note the 2020 version of EJSCREEN (confusingly released mid 2021 rather than late 2020)
    actually used ACS2018, which was from 2014-2018 (released by Census December 2019). See \link{bg20}

  Columns added here, not in ftp version:
    - lat lon for bg centroids
    - state name and state abbrev, county name, and FIPS for tract, county, state

  Detailed info on demog subgroups is in \link{bg22DemographicSubgroups2016to2020}
 }
}
\details{
This will give a quick look at some key stats:
   #round(data.frame(cbind(
   #   subgroups = unlist(ustotals(bg22DemographicSubgroups2016to2020)),
   #  maingroups = unlist(ustotals(subset(bg22_ACS2016_2020, ST != 'PR')))
   #  ), 2)

   Note Puerto Rico may not be in bg22DemographicSubgroups2016to2020,
   (but this may change)
   but EJScreen datasets do have it.

  The data could be created for this package as follows: \cr\cr
    \code{require(ejscreen); require(ACSdownload)} \cr
    \code{require(proxistat) # for adding the lat lon of each block group }\cr
    \code{require(analyze.stuff); require(ejanalysis); require(readr)} \cr
     \cr
  \code{
  #  bg <- ejscreen::ejscreen.download(yr=2022) # to be confirmed later  ------------------   \cr

    } \cr\cr
  #
  # ejscreen.download() adds fields called FIPS.TRACT, FIPS.COUNTY, FIPS.ST, countyname. \cr\cr

  # and adds lat, lon fields for block group centroids, via \link[proxistat]{bg.pts} from proxistat package\cr
  \code{# plot(bg$lon[bg$lon < -50], bg$lat[bg$lon < -50], pch = '.')} \cr\cr
  # \code{sum(is.na(bg$FIPS.ST)) # check if any are NA} \cr

  # THEN USE A NAME SPECIFIC TO THE YEAR: \cr

  \code{bg22_ACS2016_2020 <- bg; rm(bg)} \cr
  \code{save(bg22_ACS2016_2020, file = 'bg22_ACS2016_2020.rdata')} \cr \cr
}
\concept{datasets}
