% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.fieldnames.ejscreen.csv.R
\name{change.fieldnames.ejscreen.csv}
\alias{change.fieldnames.ejscreen.csv}
\title{Change colnames of csv file on EJSCREEN FTP site to nicer colnames}
\usage{
change.fieldnames.ejscreen.csv(mynames)
}
\arguments{
\item{mynames}{A character vector of colnames from a data.frame, like names(mydf). No default.}
}
\value{
Returns a character vector of colnames, same length as input parameter
}
\description{
Just a wrapper to help easily change colnames used in csv file on EJSCREEN FTP site
  into friendlier, preferred colnames for work in R. Uses \code{\link[analyze.stuff]{change.fieldnames}}
}
\examples{
 \dontrun{
 gdbtable <- ejscreen.download()
 names(gdbtable) <- change.fieldnames.ejscreen.csv(names(gdbtable))
 }
}
\seealso{
\code{\link[analyze.stuff]{change.fieldnames}}, \link{ejscreenformulas},
}
