% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.ejscreen.R
\name{download.ejscreen}
\alias{download.ejscreen}
\title{Download the EJSCREEN Dataset for use in R}
\source{
See \url{http://www.epa.gov/ejscreen} for more information, and see \url{http://www.epa.gov/ejscreen/download-ejscreen-data} or \url{ftp://newftp.epa.gov/EJSCREEN} for raw data.
}
\usage{
download.ejscreen(folder = getwd(),
  ftpurl = "ftp://newftp.epa.gov/EJSCREEN",
  zipname = "EJSCREEN_20150505.csv.zip", csvname = "EJSCREEN_20150505.csv",
  addflag = FALSE, cutoff = 80, or.tied = TRUE)
}
\arguments{
\item{folder}{Optional path to folder (directory) where the file will be downloaded and unzipped. Default is current working directory.}

\item{ftpurl}{Optional. Default is 'ftp://newftp.epa.gov/EJSCREEN' for where to find the zipped data.}

\item{zipname}{Optional. Default is 'EJSCREEN_20150505.csv.zip' for the name of the zip file of data.}

\item{csvname}{Optional. Default is 'EJSCREEN_20150505.csv' for the name of the csv file that was zipped.}

\item{addflag}{Optional. Default is FALSE. If TRUE, it adds a field called flagged, which is TRUE if 1 or more of the EJ Indexes is at/above the cutoff US percentile.}

\item{cutoff}{Optional. Default is 80. See addflag parameter.}

\item{or.tied}{Optional. Default is TRUE, meaning at or above the cutoff. FALSE means above only. See addflag parameter.}
}
\value{
Returns a data.frame with ejscreen dataset of environmental and demographics indicators, and EJ Indexes,
  as raw values, US percentiles, text for popups. Output has one row per block group.
}
\description{
Download EJSCREEN dataset from FTP site, and import to R as data.table, optionally adding a flag field.
}
\details{
Designed for 2015 version only right now. Not tested.
}
\examples{
\dontrun{

 bg <- download.ejscreen(folder='~', addflag=TRUE)
 }
}
\seealso{
\code{\link{ejscreen.create}}
}

