% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejformula.R
\name{ejformula}
\alias{ejformula}
\title{See formula(s) used for EJSCREEN variable(s)}
\usage{
ejformula(fieldname = "all", decreasing = NA, dropNA = TRUE, recursive = FALSE)
}
\arguments{
\item{fieldname}{Optional, character vector specifying variable(s) in \code{ejscreenformulas$Rfieldname}, default is all \code{ejscreenformulas$Rfieldname} that are not NA values.}

\item{decreasing}{Optional, passed to \code{\link{sort}} except default is not sorted (just the order that exists in \code{\link{ejscreenformulas}})}

\item{dropNA}{Be careful: Optional, default is TRUE. If TRUE, returns only formulas that are not NA values. If FALSE, and decreasing is not specified (sorting drops NA values here), returns vector the same length as \code{fieldname} (unless \code{recursive = TRUE})}

\item{recursive}{Optional, default is FALSE. If TRUE, returns also returns formula(s) for variable(s) found on right hand side of formula(s), i.e. those used to create specified variable(s)}
}
\value{
Character vector of the formula(s) used to calculate the specified variable, in \code{ejscreenformulas}
}
\description{
Just a convenient way to look at the formula(s) used to create one or more variables in EJSCREEN.
}
\examples{
  ejformula('VSI.eo')
  ejformula(c('pctmin', 'pctlowinc'))
  ejformula('VSI.eo', recursive = TRUE)
  ejformula()
}
\seealso{
\code{\link{ejscreenformulas}}
}
