% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ejscreen.acs.rename.R
\name{ejscreen.acs.rename}
\alias{ejscreen.acs.rename}
\title{Rename Fields of ACS Data for Use in EJSCREEN}
\usage{
ejscreen.acs.rename(acsraw, folder = getwd(), formulafile)
}
\arguments{
\item{acsraw}{Data.frame of raw data counts for each block group, such as population or number of Hispanics.}

\item{folder}{Default is getwd(). Specifies path for where to read from (if formulafile specified) and write to.}

\item{formulafile}{Default if this is blank is to use data(ejscreenformulas). Otherwise filename must be specified.
If not specified, function loads this as data().}
}
\value{
Returns a data.frame with some or all of input fields, plus calculated new fields.
}
\description{
Start with raw counts from demographic survey data, and environmental data,
  and rename fields to use friendly variable names.
}
\examples{
# (no examples yet)
}

