% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.download.R
\name{ejscreen.download}
\alias{ejscreen.download}
\title{Download the EJSCREEN Dataset csv for use in R and rename variables}
\source{
See \url{http://www.epa.gov/ejscreen} for more information, and see \url{http://www.epa.gov/ejscreen/download-ejscreen-data}
}
\usage{
ejscreen.download(
  folder = getwd(),
  yr = NULL,
  ftpurlbase = "https://gaftp.epa.gov/EJSCREEN/",
  justreadname = NULL,
  statepctiles = FALSE,
  addflag = FALSE,
  cutoff = 80,
  or.tied = TRUE
)
}
\arguments{
\item{folder}{Optional path to folder (directory) where the file will be downloaded and unzipped. Default is current working directory.}

\item{yr}{Default is latest available year found as a folder on the FTP site.
Optional numeric year designating EJSCREEN version such as 2015, 2016, 2017, 2018, 2019, 2020, etc.}

\item{ftpurlbase}{Optional. where to find the zipped data.}

\item{justreadname}{Optional character file name - if specified, skips downloading and just tries to read previously-downloaded csv found in \code{folder}.}

\item{statepctiles}{Optional, default FALSE. If TRUE, gets State Percentiles csv instead of the USPR file.}

\item{addflag}{Optional. Default is FALSE. If TRUE, it adds a field called flagged, which is TRUE if 1 or more of the EJ Indexes is at/above the cutoff US percentile.}

\item{cutoff}{Optional. Default is 80. See addflag parameter.}

\item{or.tied}{Optional. Default is TRUE, meaning at or above the cutoff. FALSE means above only. See addflag parameter.}
}
\value{
Returns a data.frame with ejscreen dataset of environmental and demographics indicators, and EJ Indexes,
  as raw values, US percentiles, text for popups. Output has one row per block group, sorted by FIPS.
}
\description{
Download EJSCREEN dataset from FTP site, unzip if necessary,
  import to R as data.table,
  renaming fields with friendly colnames,
  optionally adding a flag field (see parameter called \code{addflag}).
  Note that since 2020v, State percentiles are also available in a separate zipped csv.
}
\details{
Not fully tested. \cr
  Each version of EJSCREEN uses updated environmental data and updated 5-year summary file estimates from the American Community Survey (ACS).\cr
    \cr
  The 2021 version of EJSCREEN, likely to be released in late 2021, (which will be avail as data in \code{ejscreen::bg21}) \cr
    is based on 2015-2019 ACS (Census calls it the 2019 5-year data release, but released it in Dec 2020).\cr
    \url{https://www.census.gov/programs-surveys/acs/news/data-releases/2019/release-schedule.html}  \cr
  Note the 2020 version of EJSCREEN (confusingly released mid 2021 not late 2020)
    actually uses ACS2018, which is from 2014-2018 (released late 2019).
    It is avail as \link{bg20}
  Note the 2019 version of EJSCREEN (released late 2019)
    actually uses ACS2017, which is from 2013-2017 (released late 2018).
  Note the 2018 version of EJSCREEN (released late 2018)
    actually uses ACS2016, which is from 2012-2016 (released late 2017).
  The 2015 version of EJSCREEN, released in mid 2015, was based on 2008-2012 ACS data, and was the first public version available for download.
}
\examples{
   # bg18 <- ejscreen.download('~')
   ## bg18 <- ejscreen.download('~',
   #  justreadname = 'EJSCREEN_Full_USPR_2018.csv')
   # bg18 <- bg18[ , !grepl(pattern = 'pctile\\.text', x = names(bg18))]
   # bg18 <- bg18[ , !grepl(pattern = 'svi6', x = names(bg18))]
   # setwd('~')
   # save(bg18, file = 'bg18.rdata')
}
\seealso{
\code{\link{ejscreen.create}}  \code{\link{change.fieldnames.ejscreen.csv}}
}
