% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.rollup.all.R
\name{ejscreen.rollup.all}
\alias{ejscreen.rollup.all}
\title{Aggregate EJSCREEN Dataset at Lower Resolutions (e.g., Tracts and Counties)}
\usage{
ejscreen.rollup.all(bg, scalenames = c("tracts", "counties", "states",
  "regions"), fipsnames = c("FIPS.TRACT", "FIPS.COUNTY", "FIPS.ST", "REGION"),
  myfolder = getwd(), filenamebase = "EJSCREEN", filenames.R, filenames.csv,
  save.R = FALSE, save.csv = FALSE, assigning = FALSE, ...)
}
\arguments{
\item{bg}{Required, data.frame of raw data for environmental and demographic counts, one row per block group typically, one column per indicator.}

\item{scalenames}{optional character vector of terms used to create filenames if saving files, default = c('tracts', 'counties', 'states', 'regions')}

\item{fipsnames}{optional character vector of certain colnames in bg, used to select columns from bg to summarize by, default = c('FIPS.TRACT', 'FIPS.COUNTY', 'FIPS.ST', 'REGION'),}

\item{myfolder}{optional folder path for saving files, default = getwd()}

\item{filenamebase}{optional character element, default = 'EJSCREEN', used to construct filenames to save files if relevant.}

\item{filenames.R}{optional vector of filenames, default has the word EJSCREEN and scalename .RData}

\item{filenames.csv}{optional vector of filenames, default has the word EJSCREEN and scalename .csv}

\item{save.R}{optional logical, default = FALSE, whether to save files as .RData}

\item{save.csv}{optional logical, default = FALSE, whether to save files as .csv}

\item{assigning}{optional logical, default = FALSE, whether to assign results to variable in calling environment, or just return list of data.frames as result.}

\item{...}{Optional parameters to pass to \code{\link{ejscreen.create}} which uses formulas to create indicators from raw values.}
}
\value{
Returns a list of data.frames each like output of \code{\link{ejscreen.rollup}}, one per resolution (e.g., one for counties)
}
\description{
Does what ejscreen.rollup does, but for more than one resolution - a batch of rollups done at once. 
  Start with full EJSCREEN dataset at one resolution (typically block groups), 
  and create aggregated data at higher geographic scales (e.g., tracts and counties)
}
\details{
**default fieldnames are assumed for now. Uses \code{\link{ejscreen.create}}
}
\examples{
# (none)
}
\seealso{
\code{\link{ejscreen.rollup}}
}
