% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.rollup.R
\name{ejscreen.rollup}
\alias{ejscreen.rollup}
\title{Aggregate EJSCREEN Dataset at Lower Resolution (e.g., Tracts)}
\usage{
ejscreen.rollup(
  bg,
  fipsname = "FIPS.TRACT",
  scalename = "tracts",
  enames,
  folder = getwd(),
  sumnames,
  avgnames,
  wts,
  acsnames,
  ...
)
}
\arguments{
\item{bg}{Data.frame of raw data for environmental and demographic counts, one row per block group typically, one column per indicator.}

\item{fipsname}{Default is 'FIPS.TRACT' - specifies colname of unique ID field FIPS used to group by. Can be FIPS.TRACT, FIPS.COUNTY, FIPS.ST, or REGION in default dataset.}

\item{scalename}{***Not used. Default is 'tracts' - specifies text to use in naming the saved file.}

\item{enames}{Default is \code{\link{names.e}}, the colnames of raw envt indicators in bg}

\item{folder}{***Not used. Optional, default is getwd().}

\item{sumnames}{Default is a vector of colnames in bg, those which should be rolled up as sums (e.g., sum of all block group population counts in the tract)}

\item{avgnames}{Default is a vector of colnames in bg, those which should be rolled up as weighted averages (e.g., pop wtd mean of air pollution level)}

\item{wts}{Default is 'pop', the colname in bg specifying the field to use when calculating the weighted mean of all blockgroups in a tract, for example.}

\item{acsnames}{Not used. Default is a vector of demographic colnames in bg, used in default ejscreen dataset (see code or \code{\link{ejscreenformulas}})}

\item{...}{Optional parameters to pass to \code{\link{ejscreen.create}} which uses formulas to create indicators from raw values.}
}
\value{
Returns a data.frame with ejscreen dataset of environmental and demographics indicators, and EJ Indexes,
  as raw values, US percentiles, but not text for popups.
  *** Output has one row per tract, county, state, or region, depending on what is specified.
}
\description{
Start with full EJSCREEN dataset at one resolution (typically block groups),
  and create aggregated data at a higher geographic scale (e.g., tracts or counties)
}
\details{
**default fieldnames are assumed for now. Uses \code{\link{ejscreen.create}}
}
\examples{

 \dontrun{
  # load("~/Dropbox/EJSCREEN/R analysis/bg 2015-04-22 Rnames plus subgroups.RData")
  # Do this for each of several levels of resolution
  #
  fipsnames <- c('FIPS.TRACT', 'FIPS.COUNTY', 'FIPS.ST', 'REGION')
  scalenames <- c('tracts', 'counties', 'states', 'regions')
  # or just for tracts, say this:
  #   fipsnames <- 'FIPS.TRACT'; scalenames <- 'tracts'

  for (i in 1:length(fipsnames)) {

  ##################################### #
  # Specify resolution of interest
  fipsname <- fipsnames[i] # 'FIPS.TRACT'
  scalename <- scalenames[i] # 'tracts'

  ##################################### #
  # Get results, using the function
  myrollup <- ejscreen.rollup(bg = bg, fipsname = fipsname, scalename = scalename)

  ##################################### #
  # Save results
  save(myrollup, file = paste('EJSCREEN 2015', scalename, 'data.RData') )
  write.csv(myrollup, row.names = FALSE, file = paste('EJSCREEN 2015', scalename, 'data.csv'))

  }
 }

}
\seealso{
\code{\link{ejscreen.create}}
}
