% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.rollup.save.R
\name{ejscreen.rollup.save}
\alias{ejscreen.rollup.save}
\title{Helper for ejscreen.rollup.all, to save files of results}
\usage{
ejscreen.rollup.save(myrollup, myfolder = getwd(),
  filenamebase = "EJSCREEN", scalename = c("tracts"), filename.R,
  filename.csv, save.R = TRUE, save.csv = TRUE)
}
\arguments{
\item{myrollup}{Required, data.frame results from ejscreen.rollup.all (just one scale at a time though)}

\item{myfolder}{optional folder path for saving files, default = getwd()}

\item{filenamebase}{optional character element, default = 'EJSCREEN', used to construct filenames to save files if relevant.}

\item{scalename}{optional character term used to create filenames, default = c('tracts')}

\item{filename.R}{optional filename, default has the word EJSCREEN and scalename .RData}

\item{filename.csv}{optional filename, default has the word EJSCREEN and scalename .csv}

\item{save.R}{optional logical, default = FALSE, whether to save files as .RData}

\item{save.csv}{optional logical, default = FALSE, whether to save files as .csv}
}
\value{
Returns a 2 element vector with full paths of saved R and csv files (or NA instead of a path, if one of those is not saved)
}
\description{
Just saves csv and/or RData file(s)
}
\seealso{
\code{\link{ejscreen.rollup.all}}
}
