% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ileileplot.R
\name{ileile_plot}
\alias{ileile_plot}
\title{Scatter plot of median of Environmental score percentiles for each Demographic score percentile}
\usage{
ileile_plot(
  x_demog_percentile,
  y_envt_percentile,
  xlab = "Demographic percentile (binned by rounding)",
  ylab = "median of Environmental percentiles",
 
    main = "Does Environmental indicator tend to be higher where Demographic \% is higher?",
  ...
)
}
\arguments{
\item{x_demog_percentile}{demographic scores of block groups, as percentiles 0-100}

\item{y_envt_percentile}{environmental indicator scores of block groups, as percentiles 0-100}

\item{xlab}{optional x axis text for demographic percentile variable}

\item{ylab}{optional y axis text for environmental percentile variable}

\item{main}{optional main title for plot}

\item{...}{passed to plot(), such as col='red'}
}
\description{
Scatter plot of median of Environmental score percentiles for each Demographic score percentile
}
\examples{

ileile_plot(
  ejscreen::bg20$pctile.pctmin, 
  ejscreen::bg20$pctile.proximity.tsdf, 
  xlab='low income percentile', 
  ylab='median pctile of environmental score at that level of \% Demographics'
 )
 
 pj(1,3)
 
  
}
