% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeansByGroup_and_Ratios.R
\docType{data}
\name{MeansByGroup_and_Ratios}
\alias{MeansByGroup_and_Ratios}
\title{average of each environmental indicator in each demographic group and ratio to rest of US}
\description{
for the EJScreen 2.1 dataset, average of each environmental indicator in each demographic group and ratio to rest of US
}
\details{
see ejanalysis::RR.means() and  ejscreen/inst/SCRIPT_make_MeansByGroup_and_Ratios_RRS.US22.R
}
