% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.d.nice.R
\docType{data}
\name{names.d.nice}
\alias{names.d.nice}
\title{Nicer names for demog fields in ejscreen data}
\format{
character vector
}
\description{
This data set provides nicer names for the ejscreen demographic indicator variables.
  These can be used to label graphs, for example.
}
\details{
Defaults to the latest version
}
\seealso{
\code{\link{ejscreenformulas}} \code{\link{names.e}} \code{\link{names.d}} \code{\link{names.ej}}
}
\concept{datasets}
