% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.e.nice.R
\docType{data}
\name{names.e.nice}
\alias{names.e.nice}
\title{Nicer names for envt fields in ejscreen data}
\format{
character vector
}
\usage{
data('names.e.nice')
}
\description{
This data set provides nicer names for the ejscreen environmental indicator variables.
  These can be used to label graphs, for example.
}
\details{
Defaults to the latest version
}
\seealso{
\code{\link{ejscreenformulas}} \code{\link{names.e}}
}
\concept{datasets}
