% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.ejvars.R
\docType{data}
\name{names.ejvars}
\alias{names.ejvars}
\alias{names.ej}
\alias{EJ-variable-names}
\title{Fieldnames of environmental justice indicator columns in ejscreen package data}
\format{A series of variables (each is a character vector of colnames):
\itemize{
 \item "names.ej"
 \item "names.ej.bin"
 \item "names.ej.burden.eo"
 \item "names.ej.burden.eo.bin"
 \item "names.ej.burden.eo.pctile"
 \item "names.ej.burden.svi6"
 \item "names.ej.burden.svi6.bin"
 \item "names.ej.burden.svi6.pctile"
 \item "names.ej.pct.eo"
 \item "names.ej.pct.eo.bin"
 \item "names.ej.pct.eo.pctile"
 \item "names.ej.pct.svi6"
 \item "names.ej.pct.svi6.bin"
 \item "names.ej.pct.svi6.pctile"
 \item "names.ej.pctile"
 \item "names.ej.svi6"
 \item "names.ej.svi6.bin"
 \item "names.ej.svi6.pctile"
 \item "namesall.ej"
 \item "namesall.ej.bin"
 \item "namesall.ej.pctile"
}
And names.ej in turn is this, for example: 
\itemize{
 \item  [1] "EJ.DISPARITY.pm.eo"             
 \item  [2] "EJ.DISPARITY.o3.eo"             
 \item  [3] "EJ.DISPARITY.cancer.eo"         
 \item  [4] "EJ.DISPARITY.resp.eo"           
 \item  [5] "EJ.DISPARITY.dpm.eo"            
 \item  [6] "EJ.DISPARITY.pctpre1960.eo"     
 \item  [7] "EJ.DISPARITY.traffic.score.eo"  
 \item  [8] "EJ.DISPARITY.proximity.npl.eo"  
 \item  [9] "EJ.DISPARITY.proximity.rmp.eo"  
 \item  [10] "EJ.DISPARITY.proximity.tsdf.eo" 
 \item  [11] "EJ.DISPARITY.proximity.npdes.eo"
}}
\source{
Names developed for this package. No external data source.
}
\usage{
data('names.ejvars')
}
\description{
This data set provides variables that hold the colnames
  of environmental indicator fields in data.frames that may be used in the ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.ej]
}
\details{
This should have the latest (2016) version. Also see \code{\link{names.ejvars16}}). 
 The 2015 version had neuro-related indicators in it, and is now in \code{\link{names.ejvars15}}.
}
\seealso{
\code{\link{ejscreenformulas}} \code{\link{names.evars}} \code{\link{names.dvars}} \code{\link{names.ejvars}}
}
\concept{datasets}
