% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.evars15.R
\docType{data}
\name{names.evars15}
\alias{Elist}
\alias{environmental-variable-names}
\alias{names.e}
\alias{names.evars}
\alias{names.evars15}
\title{2015 Fieldnames of environmental indicator columns in ejscreen package data}
\format{A series of variables (each is a character vector of colnames).
For 2016 version of EJSCREEN:
\itemize{
 \item "names.e" (pm, o3, cancer,     resp, dpm, pctpre1960, traffic.score, proximity.npl, proximity.rmp, proximity.tsdf, proximity.npdes)
 \item "names.e.bin"
 \item "names.e.pctile"
 \item "Elist" (this one is like names.e, but as a list, not a vector)
}
For 2015 version of EJSCREEN it was changed so names include 15, to distinguish:
\itemize{
 \item "names.e15" (pm, o3, cancer, neuro, resp, dpm, pctpre1960, traffic.score, proximity.npl, proximity.rmp, proximity.tsdf, proximity.npdes)
 \item "names.e.bin15"
 \item "names.e.pctile15"
 \item "Elist15" (this one is like names.e, but as a list, not a vector)
}}
\source{
Names developed for this package. No external data source.
}
\usage{
data('names.evars')
}
\description{
This data set provides variables that hold the colnames
  of environmental indicator fields in data.frames that may be used in the ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.e15]
}
\details{
NOTE: This is to provide the 2015 version's list, which had "neuro" in it, but now defaults to the latest version (e.g., 2016 or whatever).
}
\seealso{
\code{\link{ejscreenformulas}} \code{\link{names.evars}} \code{\link{names.dvars}} \code{\link{names.ejvars}}
}
\keyword{datasets}

