% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.evars16.R
\docType{data}
\name{names.evars16}
\alias{names.evars16}
\alias{names.evars}
\alias{names.e}
\alias{environmental-variable-names}
\alias{Elist}
\title{Fieldnames of environmental indicator columns in ejscreen package data}
\format{A series of variables (each is a character vector of colnames).
For 2016 version of EJSCREEN:
\itemize{
 \item "names.e" (pm, o3, cancer,     resp, dpm, pctpre1960, traffic.score, proximity.npl, proximity.rmp, proximity.tsdf, proximity.npdes)
 \item "names.e.bin"
 \item "names.e.pctile"
 \item "Elist" (this one is like names.e, but as a list, not a vector)
}
For 2015 version of EJSCREEN it was:
\itemize{
 \item "names.e" (pm, o3, cancer, neuro, resp, dpm, pctpre1960, traffic.score, proximity.npl, proximity.rmp, proximity.tsdf, proximity.npdes)
 \item "names.e.bin"
 \item "names.e.pctile"
 \item "Elist" (this one is like names.e, but as a list, not a vector)
}}
\source{
Names developed for this package. No external data source.
}
\usage{
data('names.evars')
}
\description{
This data set provides variables that hold the colnames
  of environmental indicator fields in data.frames that may be used in the ejscreen package
  to make it easier to refer to them as a vector, e.g., mydf[ , names.e]
}
\details{
This is the 2016 version
}
\seealso{
\code{\link{names.e.nice}}  \code{\link{ejscreenformulas}} \code{\link{names.dvars}} \code{\link{names.ejvars}}
}
\keyword{datasets}
