% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenames.R
\name{nicenames}
\alias{nicenames}
\title{Convert EJSCREEN R variable names to more descriptive labels from glossary}
\usage{
nicenames(x)
}
\arguments{
\item{x}{EJSCREEN variable names as found in ejscreenformulas$Rfieldnames, one or more as character vector}
}
\value{
character vector same shape as x, with nice name if avail or unchanged x where otherwise
}
\description{
Provides more descriptive variable names drawn from the {\link{ejscreenformulas}},
using the ejscreenformulas$glossaryfieldname that corresponds to each ejscreenformulas$Rfieldname in x
For example it provides "% low-income (i.e., with income below 2 times poverty level)"
if given "pctlowinc"
}
\examples{
 nicenames(names.e)
 nicenames(names.d)
 nicenames(names.d.subgroups)
}
