% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctileAsText.R
\name{pctileAsText}
\alias{pctileAsText}
\title{Utility function in showing a percentile as popup text}
\usage{
pctileAsText(x)
}
\arguments{
\item{x}{vector or data.frame of numeric values 0 to 100 (not 0 to 1), 
representing percentiles from EJSCREEN dataset}
}
\value{
Returns matrix/vector of same shape as x if x was data.frame/vector
}
\description{
Converts numeric percentiles (0-100) into character (text)
  that converts 95.3124 to '95%ile' for example.
}
\examples{
 \dontrun{
 (bg2[ 125:126, c('pctile.pctmin', 'pctile.EJ.DISPARITY.pm.eo') ])
 (bg2[ 125:126, c('pctile.text.pctmin', 'pctile.text.EJ.DISPARITY.pm.eo') ])
 pctileAsText(bg2[ 125:126, c('pctile.pctmin', 'pctile.EJ.DISPARITY.pm.eo') ])
 }
}
