% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\name{pj}
\alias{pj}
\title{plot ej percentiles vs percentiles from EJSCREEN data}
\usage{
pj(enum, dnum, dat = ejscreen::bg20, ...)
}
\arguments{
\item{enum}{which of the envt indicators to use, such as 1}

\item{dnum}{which of the demog indicators to use, such as 3}

\item{dat}{data.frame of ejscreen data, default is bg20}
}
\description{
plot ej percentiles vs percentiles from EJSCREEN data
}
\examples{
  pj(1,3)
  pj(1,6, col='blue', main='My Graphic')

}
