% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.US.R
\docType{data}
\name{RRS.US}
\alias{RRS.US}
\title{Ratios of mean indicator values across demographic groups}
\description{
Based on the latest EJSCREEN dataset

    See \code{\link[ejanalysis]{RR.table}} for how this was created and can be used.
 \preformatted{
    bg <- ejscreen::bg21plus # or ejscreen::bg21 ?
    names.dplus <- c(ejscreen::names.d, ejscreen::names.d.subgroups.pct)

    RRS.US21 <- ejanalysis::RR.table(
      mydat =  bg,
      Enames = ejscreen::names.e,
      Dnames = names.dplus,
      popcolname = 'pop'
    )
 }
}
