% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg21.R
\docType{data}
\name{bg21}
\alias{bg21}
\title{The (late) 2021 version of EJSCREEN data (based on ACS 2015-2019) plus lat lon, countynames, etc., minus some nonessential fields}
\format{
data.frame with 220,333 ? rows (block groups) and 118 ? columns in bg21
}
\description{
\preformatted{
  The 2021 version of EJSCREEN, likely to be released in late 2021, (which will be avail as data in \code{ejscreen::bg21}) \cr
    is based on 2015-2019 ACS (Census calls it the 2019 5-year data release, but released it in Dec 2020).\cr
    \url{https://www.census.gov/programs-surveys/acs/news/data-releases/2019/release-schedule.html}  \cr

  Note the 2020 version of EJSCREEN (confusingly released mid 2021 rather than late 2020)
    actually uses ACS2018, which is from 2014-2018 (released by Census late 2019). See \link{bg20}

  Note the 2019 version of EJSCREEN (released late 2019)
    actually uses ACS2017, which is from 2013-2017 (released by Census late 2018).

  This data set is the EJSCREEN dataset from the ftp site but with
  fields renamed for easier use in the ejscreen package,
  Columns dropped from ftp version, not included here:
    - text fields about percentiles, used as popup text on maps
    - any obsolete columns related to the 2 alternative versions of an EJ Index phased out (except VNI.eo and VSI.eo are kept)
    - any obsolete columns related to using 6 instead of 2 demographic variables (svi6)
  Columns added here, not in ftp version:
    - lat lon for bg centroids
    - state name and state abbrev, county name, and FIPS for tract, county, state
    - flag indicating if any of EJ indexes above 80th percentile in US

  Detailed info on demog subgroups is in \link{bg21DemographicSubgroups2015to2019}
 }
}
\details{
#'   This will give a quick look at some key stats:
   round(data.frame(cbind(
     subgroups=unlist(ustotals(bg21DemographicSubgroups2015to2019)),
     maingroups= unlist(ustotals(bg21[bg21$ST !='PR',])))
     ),2)

Note that unlike bg21DemographicSubgroups2015to2019,
   bg21 has PR, but will likely add PR to have it in both.

  Lookup tables are not yet included here: lookup tables of percentiles for USA, Regions, States are on EJSCREEN FTP site.
  For that information, see ejscreen::lookupUSA, lookupRegions, lookupStates, and ejscreen::ejscreen.lookuptables

  Definitions of column names here vs in the gdb or csv on the FTP site files
  are in data(ejscreenformulas)
  and also in a file called EJSCREEN_columns_explained.csv
  in the inst folder of the ejscreen package.
  Technical documentation is at \url{https://www.epa.gov/ejscreen}

  The data was created for this package as follows: \cr\cr
    \code{require(ejscreen); require(ACSdownload)} \cr
    \code{require(proxistat) # for adding the lat lon of each block group }\cr
    \code{require(analyze.stuff); require(ejanalysis); require(readr)} \cr
     \cr
  \code{ # DOWNLOADED AND UNZIPPED  }\cr\cr
  \cr Latest version of EJSCREEN downloaded from public FTP site
  \cr as csv format (zipped)
  \cr then unzipped to csv file
  \cr
  \code{
  #  bg <- ejscreen.download(folder = getwd(), addflag = TRUE,   \cr
   justreadname = 'EJSCREEN_2021_USPR.csv' )} \cr\cr
  # Starts by reading in approx 141 columns from csv\cr
  # and \link{ejscreen.download} renames fields, drops an ID field, \cr
  # and adds fields called FIPS.TRACT, FIPS.COUNTY, FIPS.ST, countyname, and (optionally) flagged. \cr\cr
  \code{bg <- bg[ , !grepl(pattern = 'pctile\\\\.text', x = names(bg))]} \cr
  \code{bg <- bg[ , names(bg) != 'Shape_Length'] } \cr

  # Then added lat, lon fields for block group centroids, via \link[proxistat]{bg.pts} from proxistat package:\cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lat')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr
  \code{bg <- merge(bg, proxistat::bg.pts[ , c('FIPS', 'lon')], by.x = 'FIPS', by.y = 'FIPS', all.x = TRUE, all.y = FALSE)} \cr\cr
  \code{plot(bg$lon[bg$lon < -50], bg$lat[bg$lon < -50], pch = '.')} \cr\cr
  # \code{sum(is.na(bg$FIPS.ST)) # check if any are NA} \cr

  # THEN USE A NAME SPECIFIC TO THE YEAR: \cr
  \code{bg21 <- bg; rm(bg)} \cr
  \code{save(bg21, file = 'bg21.rdata')} \cr \cr
}
\concept{datasets}
