% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ejscreen.create.R
\name{ejscreen.create}
\alias{ejscreen.create}
\title{Create EJSCREEN Dataset from Environmental Indicators}
\usage{
ejscreen.create(bg.e, acsraw, folder = getwd())
}
\arguments{
\item{bg.e}{Data.frame of raw data for environmental indicators, one row per block group, one column per indicator.}

\item{acsraw}{Data.frame of raw demographic indicators.}

\item{folder}{Where to save results. Default is getwd()}
}
\value{
Returns a data.frame with full ejscreen dataset of environmental and demographics indicators, and EJ Indexes,
  as raw values, US percentiles, and text for popups. Output has one row per block group.
}
\description{
Start with raw environmental indicator data, and create full EJSCREEN dataset.
  This code also contains an outline of steps involved.
}
\examples{
\dontrun{
 mybg.e=data.frame(FIPS=analyze.stuff::lead.zeroes(1:1000, 12), air=rlnorm(1000), water=rlnorm(1000)*5, stringsAsFactors=FALSE)
 x=ejscreen.create(envirodata) # downloads ACS demographics and combines with user provided envirodata
 # does not work like this since needs more Demog vars:
 # myacsraw=data.frame(FIPS=analyze.stuff::lead.zeroes(1:1000, 12), pop=rnorm(n=1000, mean=1400, sd=200), mins=runif(1000, 0, 800), lowinc=runif(1000, 0,500), stringsAsFactors=FALSE)
 # y=ejscreen.create(bg.e=mybg.e, acsraw=myacsraw)
 }
}

