% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.create.R
\name{ejscreen.create}
\alias{ejscreen.create}
\title{Create EJSCREEN Dataset from Environmental Indicators}
\usage{
ejscreen.create(e, acsraw, folder = getwd(), keep.old, formulas,
  mystates = "all", demogvarname0 = "VSI.eo",
  demogvarname1 = "VSI.svi6", wtsvarname = "pop", checkfips = TRUE,
  EJprefix0 = "EJ.DISPARITY", EJprefix1 = "EJ.BURDEN",
  EJprefix2 = "EJ.PCT", ejformulasfromcode = FALSE, ejtype = 1,
  demogvarname0suffix = "eo", demogvarname1suffix = "svi6", end.year,
  threshold = FALSE, cutoff = 0.8, thresholdfieldnames, ...)
}
\arguments{
\item{e}{Data.frame of raw data for environmental indicators, one row per block group, one column per indicator.}

\item{acsraw}{Optional data.frame of raw demographic indicators. Downloaded if not provided as parameter.}

\item{folder}{Optional, default is getwd(). Passed to \code{\link[ACSdownload]{get.acs}} if demog data must be downloaded.
Passed to but not currently used by ejscreen.acs.rename which uses \code{\link[analyze.stuff]{change.fieldnames}} in \pkg{analyze.stuff} package.
Not currently passed to ejscreen.acs.calc which uses \code{\link[analyze.stuff]{calc.fields}} in \pkg{analyze.stuff} package.}

\item{keep.old}{optional vector of colnames from e that are to be used/returned. For nondefault colnames, this must be used.}

\item{formulas}{optional, see \code{\link{ejscreen.acs.calc}} for details. Defaults are in ejscreenformulas$formula
Note that if formulas is specified, ejformulasfromcode is ignored.}

\item{mystates}{optional vector of 2-letter state abbreviations. Default is "all" which specifies all states plus DC (BUT NOT PR - we exclude PR so that calculating US percentiles works right)}

\item{demogvarname0}{optional, default is 'VSI.eo' used as demographic indicator for EJ Indexes. Must be a colname in acsraw or created and kept by formulas.}

\item{demogvarname1}{optional, default is 'VSI.svi6' used for alternative EJ Indexes. Must be a colname in acsraw or created and kept by formulas.}

\item{wtsvarname}{optional, default is 'pop' used for weighted percentiles, etc. Must be a colname in acsraw or created and kept by formulas.}

\item{checkfips}{optional, default is TRUE. If TRUE, function checks to verify all FIPS codes appear to be valid US FIPS
(correct number of characters, adding any leading zero needed, and checking the first five to ensure valid county). To use something other than actual US FIPS codes, set this to FALSE.}

\item{EJprefix0}{optional, default is 'EJ.DISPARITY' - specifies prefix for colnames of main EJ Indexes, with a period separating prefix from body of colname}

\item{EJprefix1}{optional, default is 'EJ.BURDEN' - specifies prefix for colnames of Alternative 1 version of EJ Indexes, with a period separating prefix from body of colname}

\item{EJprefix2}{optional, default is 'EJ.PCT' - specifies prefix for colnames of Alternative 2 version of EJ Indexes, with a period separating prefix from body of colname}

\item{ejformulasfromcode}{optional, default is FALSE. If TRUE, use EJ Index formulas built into this function instead of the EJ Index formulas in ejscreenformulas.
The parameters such as demogvarname0 are only used if ejformulasfromcode=TRUE. Note that if formulas is specified, ejformulasfromcode is ignored.}

\item{ejtype}{optional, default is 1, defines which formula to use for ejindex if not using ejscreenformulas. See \code{\link{ej.indexes}} But note alt1 and alt2 still use type 5 and 6 ignoring ejtype.}

\item{demogvarname0suffix}{optional, default is 'eo' - specifies suffix for colnames of EJ Indexes based on demogvarname0, with a period separating body of colname from suffix}

\item{demogvarname1suffix}{optional, default is 'svi6' - specifies suffix for colnames of EJ Indexes based on demogvarname1, with a period separating body of colname from suffix}

\item{end.year}{optional to pass to \code{\link[ACSdownload]{get.acs}} (such as end.year='2013' -- otherwise uses default year used by \code{\link[ACSdownload]{get.acs}})}

\item{threshold}{optional, default is FALSE. Set to TRUE to add a column (called 'flagged') to results that is TRUE when one or more of certain percentiles (US EJ Index) in a block group (row) exceed cutoff.
A field called flagged can also be added via \link[ejanalysis]{flagged} ejanalysis::flagged() or via \link{ejscreen.download}( addflag = TRUE )}

\item{cutoff}{optional, default is 0.80 (80th percentile). If threshold=TRUE, then cutoff defines the threshold against which percentiles are compared.}

\item{thresholdfieldnames}{optional, default is standard EJSCREEN EJ Indexes built into code. Otherwise, vector of character class fieldnames, specifying which fields to compare to cutoff if threshold=TRUE.}

\item{...}{optional extra parameters passed only to \code{\link[ACSdownload]{get.acs}} such as new.geo = FALSE, save.files = TRUE, write.files = TRUE}
}
\value{
Returns a data.frame with full ejscreen dataset of environmental and demographics indicators, and EJ Indexes,
  as raw values, US percentiles, and text for popups. Output has one row per block group.
}
\description{
Start with raw environmental indicator data, and create (or replicate) a full EJSCREEN dataset.
  The source code also contains comments with an outline of steps involved.
  Note that rather than using this function, one can instead
  download the gdb file from the EJSCREEN FTP site
  and open it in ESRI ArcGIS and export the attribute tables as text files in csv format.
}
\details{
**Note that if non-default fieldnames are used in e and/or acsraw,
  those must be specified in parameters including demogvarname0, demogvarname1, wtsvarname,
  keep.old (and could be reflected in prefix and suffix params as well).
}
\examples{
 \dontrun{
 set.seed(99)
 envirodata=data.frame(FIPS=analyze.stuff::lead.zeroes(1:1000, 12),
   air=rlnorm(1000), water=rlnorm(1000)*5, stringsAsFactors=FALSE)
 demogdata=data.frame(FIPS=analyze.stuff::lead.zeroes(1:1000, 12),
   pop=rnorm(n=1000, mean=1400, sd=200), mins=runif(1000, 0, 800),
   num2pov=runif(1000, 0,500), stringsAsFactors=FALSE)
 demogdata$povknownratio <- demogdata$pop
 # downloads ACS demographics and combines with user provided envirodata:
 # bg1=ejscreen.create(envirodata, mystates=c('de','dc'))
 # currently does not work for nonstandard colnames
 #  unless keep.old used as follows (work in progress):
 y=ejscreen.create(e=envirodata, acsraw=demogdata,
   keep.old = c(names(envirodata), names(demogdata)),
   demogvarname0 = 'pctmin', demogvarname1 = 'pctlowinc', wtsvarname = 'pop' )
 }
}
\seealso{
\code{\link{make.popup.d}} \code{\link{make.popup.e}} \code{\link{make.popup.ej}} \code{\link{ejscreen.lookuptables}}
}
