% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ejscreen.lookuptables.R
\name{ejscreen.lookuptables}
\alias{ejscreen.lookuptables}
\title{Create EJSCREEN Lookup Tables of Pop. Percentiles by Zone}
\usage{
ejscreen.lookuptables(x, weights, zone, folder = getwd(),
  missingcode = -9999999)
}
\arguments{
\item{x}{Data.frame of indicators, one row per block group, one column per indicator.}

\item{weights}{Weights for percentiles -- Default is population count to provide population percentiles.}

\item{zone}{NOT IMPLEMENTED HERE - HELPER FUNCTIONS ASSUME STATES AND REGIONS ARE ZONES NEEDED}

\item{folder}{Default is getwd() - specifies where to save the csv files.}

\item{missingcode}{Leave this unspecified if missing values are set to NA in the input data. Default is -9999999 (but if already NA then do not specify anything for this). The number or value in the input data that designates a missing value.}
}
\value{
Creates lookup tables saved as csv files to specified folder.
}
\description{
Start with raw environmental, demographic, and EJ indicator data, and write as csv files to disk a series of
  lookup tables that show population percentiles and mean values for each indicator.
}
\examples{
# (no examples yet)
}

