% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreensignifarray.R
\name{ejscreensignifarray}
\alias{ejscreensignifarray}
\title{Specify Significant Digits for Each Column of EJSCREEN Indicators}
\usage{
ejscreensignifarray(dat, digits = "ejscreen")
}
\arguments{
\item{dat}{Required, matrix or numeric data.frame with the values to be rounded.}

\item{digits}{Optional, 'ejscreen' by default. Can be a vector as long as the number of columns in dat, where each elements specifies the
number of significant digits to retain for numbers in the corresponding column of dat.
If 'ejscreen' it specifies using the default settings described below in details, 
in which case all colnames(dat) must be among (but in any order) defaultcolnames below.}
}
\value{
Returns dat, but with numbers rounded based on digits parameter.
}
\description{
Given a matrix or numeric data.frame, round each column to a specified column-specific number of significant digits.
 This function provides default values significant digits to use for an EJSCREEN environmental dataset.
 This is a wrapper for analyze.stuff::signifarray which is a wrapper that applies signif() to a matrix or data.frame.
}
\details{
Sig figs used if digits specified as 'ejscreen' are those stored in data(esigfigs)
}
\examples{
 ejscreensignifarray(data.frame(a=rnorm(10), b=rnorm(10), c=rnorm(10)), 1:3)
 envirodata <- data.frame(matrix(rnorm(11*10), ncol=11))
   # data("names.evars"); names(envirodata) <- names.e
   names(envirodata) <- c("pm", "o3", "cancer", "resp", "dpm", "pctpre1960", "traffic.score", 
   "proximity.npl", "proximity.rmp", "proximity.tsdf", "proximity.npdes")
 ejscreensignifarray(envirodata)
}
\seealso{
\code{\link{esigfigs}} \code{\link[analyze.stuff]{signifarray}} \code{\link{signif}}
}
