% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ileileplot.R
\name{ileile_plot}
\alias{ileile_plot}
\alias{qq}
\alias{Shows}
\alias{median}
\alias{envt}
\alias{percentile}
\alias{of}
\alias{places}
\alias{within}
\alias{each}
\alias{demog}
\title{Scatter plot of median of Environmental score percentiles for each Demographic score percentile
Does Environmental score tend to be higher where Demographic percent is higher?}
\usage{
ileile_plot(
  x_demog_percentile,
  y_envt_percentile,
  xlab = "Demographic percentile (binned by rounding)",
  ylab = "median of Environmental percentiles",
  main = "Does Environmental indicator tend to be higher where Demographic \% is higher?",
  ...
)
}
\arguments{
\item{x_demog_percentile}{demographic scores of block groups, as percentiles 0-100}

\item{y_envt_percentile}{environmental indicator scores of block groups, as percentiles 0-100}

\item{xlab}{optional x axis text for demographic percentile variable}

\item{ylab}{optional y axis text for environmental percentile variable}

\item{main}{optional main title for plot}

\item{...}{passed to plot(), such as col='red'}
}
\description{
Scatter plot of median of Environmental score percentiles for each Demographic score percentile
Does Environmental score tend to be higher where Demographic percent is higher?
}
\examples{
bg <- bg22
ileile_plot(
  ejscreen::bg$pctile.pctmin,
  ejscreen::bg$pctile.proximity.tsdf,
  xlab='low income percentile',
  ylab='median pctile of environmental score at that level of \% Demographics'
 )
 # also see 
 # cars::qqPlot(  something_like_pctile.proximity.rmp_at_some_sites, main="How close to US distribution of scores are scores at these sites?", ylab="RMP proximity scores at these sites (as US \%ile)", xlab = "Normal quantiles")

 pj(1,3)


}
