% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.county.R
\docType{data}
\name{RRS.county}
\alias{RRS.county}
\title{Ratios of mean indicator values across demographic groups by US County}
\description{
Based on the latest EJSCREEN dataset such as bg20

    See \code{\link[ejanalysis]{RR.table}} for how this was created and can be used.
}
