% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.ST.R
\docType{data}
\name{RRS.ST}
\alias{RRS.ST}
\title{Ratios of mean indicator values across demographic groups by State}
\description{
Based on the latest EJSCREEN dataset

    See \code{\link[ejanalysis]{RR.table}} for how this was created and can be used.
}
