% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.US.R
\docType{data}
\name{RRS.US}
\alias{RRS.US}
\title{Ratios of mean indicator values across demographic groups}
\description{
Based on the latest EJSCREEN dataset
    See   ejanalysis::RR.means() and  
    ejscreen/inst/SCRIPT_make_MeansByGroup_and_Ratios_RRS.US22.R 
    and  \code{\link[ejanalysis]{RR.table}} for how this was created and can be used.
}
