% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRS.US.R
\docType{data}
\name{RRS.US}
\alias{RRS.US}
\title{Ratios of mean indicator values across demographic groups}
\description{
Note the 2020 version of EJSCREEN (released January 2021 rather than late 2020)
    actually uses ACS2018, which is from 2014-2018 (released by Census late 2019).

    See \code{\link[ejanalysis]{RR.table}} for how this was created and can be used.
}
